+++ app-backend/api/src/main/scala/annotation-project/AnnotationProjectLabelClassRoutes.scala
@@ -1,28 +1,31 @@
 package com.rasterfoundry.api.annotationProject
 
 import com.rasterfoundry.akkautil._
+import com.rasterfoundry.api.CommonLabelClassRoutes
 import com.rasterfoundry.database._
 import com.rasterfoundry.datamodel._
 
-import akka.http.scaladsl.model._
 import akka.http.scaladsl.server._
 import cats.effect._
-import cats.implicits._
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import doobie._
 import doobie.implicits._
-import doobie.util.transactor.Transactor
 
 import java.util.UUID
 
-trait LabelClassRoutes
+trait AnnotationProjectLabelClassRoutes
     extends CommonHandlers
     with Directives
     with Authentication {
 
   val xa: Transactor[IO]
 
-  def listGroupLabelClasses(projectId: UUID, labelClassGroupId: UUID): Route =
+  def commonLabelClassRoutes: CommonLabelClassRoutes
+
+  def listAnnotationProjectGroupLabelClasses(
+      projectId: UUID,
+      labelClassGroupId: UUID
+  ): Route =
     authenticate { user =>
       authorizeScope(
         ScopedAction(Domain.AnnotationProjects, Action.Read, None),
@@ -39,17 +42,15 @@ trait LabelClassRoutes
             .transact(xa)
             .unsafeToFuture
         } {
-          complete {
-            AnnotationLabelClassDao
-              .listAnnotationLabelClassByGroupId(labelClassGroupId)
-              .transact(xa)
-              .unsafeToFuture
-          }
+          commonLabelClassRoutes.listLabelClasses(labelClassGroupId)
         }
       }
     }
 
-  def addLabelClassToGroup(projectId: UUID, groupId: UUID): Route =
+  def addAnnotationProjectLabelClassToGroup(
+      projectId: UUID,
+      groupId: UUID
+  ): Route =
     authenticate { user =>
       authorizeScope(
         ScopedAction(Domain.AnnotationProjects, Action.Update, None),
@@ -67,30 +68,16 @@ trait LabelClassRoutes
             .unsafeToFuture
         } {
           entity(as[AnnotationLabelClass.Create]) { labelClassCreate =>
-            complete(
-              StatusCodes.Created,
-              (for {
-                annotationLabelGroupOpt <- AnnotationLabelClassGroupDao
-                  .getGroupWithClassesById(groupId)
-                insert <- annotationLabelGroupOpt traverse { groupWithClass =>
-                  AnnotationLabelClassDao
-                    .insertAnnotationLabelClass(
-                      labelClassCreate,
-                      groupWithClass.toClassGroup,
-                      None
-                    )
-                }
-
-              } yield insert)
-                .transact(xa)
-                .unsafeToFuture
-            )
+            commonLabelClassRoutes.addLabelClass(labelClassCreate, groupId)
           }
         }
       }
     }
 
-  def getLabelClass(projectId: UUID, labelClassId: UUID): Route =
+  def getAnnotationProjectLabelClass(
+      projectId: UUID,
+      labelClassId: UUID
+  ): Route =
     authenticate { user =>
       authorizeScope(
         ScopedAction(Domain.AnnotationProjects, Action.Read, None),
@@ -108,18 +95,18 @@ trait LabelClassRoutes
               .transact(xa)
               .unsafeToFuture
           } {
-            complete {
-              AnnotationLabelClassDao
-                .getById(labelClassId)
-                .transact(xa)
-                .unsafeToFuture
-            }
+            commonLabelClassRoutes.getLabelClass(
+              labelClassId
+            )
           }
         }
       }
     }
 
-  def updateLabelClass(projectId: UUID, labelClassId: UUID): Route =
+  def updateAnnotationProjectLabelClass(
+      projectId: UUID,
+      labelClassId: UUID
+  ): Route =
     authenticate { user =>
       authorizeScope(
         ScopedAction(Domain.AnnotationProjects, Action.Update, None),
@@ -138,52 +125,17 @@ trait LabelClassRoutes
               .unsafeToFuture
           } {
             entity(as[AnnotationLabelClass]) { updatedClass =>
-              onSuccess(
-                AnnotationLabelClassDao
-                  .update(
-                    labelClassId,
-                    updatedClass
-                  )
-                  .transact(xa)
-                  .unsafeToFuture
-              ) {
-                completeSingleOrNotFound
-              }
+              commonLabelClassRoutes.updateLabelClass(
+                updatedClass,
+                labelClassId
+              )
             }
           }
         }
       }
     }
 
-  def activateLabelClass(projectId: UUID, labelClassId: UUID): Route =
-    authenticate { user =>
-      authorizeScope(
-        ScopedAction(Domain.AnnotationProjects, Action.Update, None),
-        user
-      ) {
-        authorizeAuthResultAsync {
-          AnnotationProjectDao
-            .authorized(
-              user,
-              ObjectType.AnnotationProject,
-              projectId,
-              ActionType.Edit
-            )
-            .transact(xa)
-            .unsafeToFuture
-        } {
-          complete {
-            AnnotationLabelClassDao
-              .activate(labelClassId)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-
-      }
-    }
-
-  def deactivateLabelClass(
+  def activateAnnotationProjectLabelClass(
       projectId: UUID,
       labelClassId: UUID
   ): Route =
@@ -203,14 +155,33 @@ trait LabelClassRoutes
             .transact(xa)
             .unsafeToFuture
         } {
-          onSuccess(
-            AnnotationLabelClassDao
-              .deactivate(labelClassId)
-              .transact(xa)
-              .unsafeToFuture
-          ) {
-            completeSingleOrNotFound
-          }
+          commonLabelClassRoutes.activeLabelClass(labelClassId)
+        }
+
+      }
+    }
+
+  def deactivateAnnotationProjectLabelClass(
+      projectId: UUID,
+      labelClassId: UUID
+  ): Route =
+    authenticate { user =>
+      authorizeScope(
+        ScopedAction(Domain.AnnotationProjects, Action.Update, None),
+        user
+      ) {
+        authorizeAuthResultAsync {
+          AnnotationProjectDao
+            .authorized(
+              user,
+              ObjectType.AnnotationProject,
+              projectId,
+              ActionType.Edit
+            )
+            .transact(xa)
+            .unsafeToFuture
+        } {
+          commonLabelClassRoutes.deactivateLabelClass(labelClassId)
         }
       }
     }